requirements.txt
flask
numpy
tensorflow
keras
opencv-python
pillow